Fix ODP table (1a, 1b) rendering
Resolves #1874
Fix inconsistency in ODP Tables 1a and 1b:
Issue:
When adding a new original data point for year before 1990 the UI broke: first 'odp' year showed as 1990, even though selected year was 1985.
Fix:
Updated the check to see if current row is included in headers, and not if first of both equal

Includes minor refactor where Table.tsx thead and tbody are separated in their own files
