+++ src/client/pages/AssessmentSection/DataTable/Table/TableBody/index.ts
@@ -1,0 +1,1 @@
+export { default } from './TableBody'
