+++ src/client/pages/AssessmentSection/DataTable/Table/TableHead/index.ts
@@ -1,0 +1,1 @@
+export { default } from './TableHead'
